Move Kotlin method name test to appropriate class
